refactor: вынести логику Telegram в подпакеты
