+++ pkg/telegram/reaction/reaction_test.go
@@ -1,4 +1,4 @@
-package telegram
+package reaction
 
 import (
 	"testing"
